Create README.md
This advanced encryption tool is developed for educational and awareness purposes and I am not responsible for any misuse in illegal ways. It provides features like BypassUAC, StartupScheduler WDexclusion, Anti-VM and Antikill. The tool ensures file protection and effectively bypasses security measures. It is intended for educational and research purposes only. The creators disclaim any responsibility for misuse, and users are fully responsible for complying with laws and regulations...